+++ src/init.c
@@ -1,0 +1,10 @@
+// src/init.c
+#include <R.h>
+#include <Rinternals.h>
+#include <R_ext/Rdynload.h>
+
+void R_init_mcGLM(DllInfo *dll)
+{
+  R_registerRoutines(dll, NULL, NULL, NULL, NULL);
+  R_useDynamicSymbols(dll, TRUE);
+}